News now uses database, other stuff
hella rad!!!
The main thing is that instead of storing stuff in a dorky in-memory
adapter, it's now stored in a SQLite database (NewsDbHelper) with a
simple content provider (NewsProvider).
In the UI, the RecyclerView has been replaced ( :( ) with a ListView,
CursorAdapter, and LoaderManager to read stuff from the DB/provider and
automatically update it. I'll see you again someday, RecyclerView.
The AsyncTask has been moved out into its own class and it writes
straight to the DB.
Now there's a WebView instead in NewsDetailActivity, which is really eh.
Oh well. (Which means ListTagHandler is out, never worked anyway.)
But that NewsDetailActivity also gets "Share" and "Open in browser"
actions, because I wanted to celebrate! Um, yeah. Cool.
And other changes as usual: IodineNewsEntry > NewsEntry (no longer needs
to be Serializable for goodness' sake), added Utils class, deleted
activity_news.xml which was unused.
(By the way, all of this was done last night. I was just too lazy to
write a commit message then.)