News now uses database, other stuff
hella rad!!!
The main thing is that instead of storing news entries in-memory, it's now stored in a SQLite database (NewsDbHelper) with a basic content provider (NewsProvider).
In the UI, the RecyclerView has been replaced with a ListView, CursorAdapter, and LoaderManager to read stuff from the DB/provider and automatically update it. (see you again someday, RecyclerView)
The AsyncTask has been moved out into its own class and it writes straight to the DB.
Now there's a WebView instead in NewsDetailActivity, which is really eh. Oh well. (Which means ListTagHandler is out, never worked anyway.)
But that NewsDetailActivity also gets "Share" and "Open in browser" actions, because I wanted to celebrate! Um, yeah. Cool.
And other cleanup as usual. (btw, this was all done last night. I was just too lazy to write a commit message then.)