Use database in ScheduleFragment
A lot of changes! The database/content provider stuff actually sort of
works now, for one. And ScheduleFragment is hooked up to that with a
CursorLoader and ListView, replacing the previous RecyclerView. This is
similar to what was done before with NewsFragment.
More random changes in model classes and parser, the latter of which is
now infinitely screwed up.  Nullability continues to be really annoying,
what am I even doing?
There's a new table, Schedule, which holds the currently selected AID
for each BID. Somehow that gets joined in with the Blocks table and
stuff.