News batch operations combined
All the batch/merging/updating stuff (there are currently 3 of them) are
now combined into one fugly method in Utils.
Easily the stupidest thing I've ever done that's somehow worked on the
first try. (Type-safe languages, tho.) And that's a good thing - the
main point is to reduce errors and code duplication, even if this is a
bit excessive.