Sync changes
Changed the way that enforcing login is handled a bit; mainly an
expedited sync is requested as soon as an account is added, which didn't
happen before, unfortunately. Also, the loaders are initialized
regardless of login, so that when sync happens after login, stuff
actually shows up.
Accidental extraneous logging of timezones in NewsListParser removed,
oops.