Use support library for Fragment, Loader, etc
Because FragmentStatePagerAdapter is only available in the support
library, that means all instances of the Fragment class
(android.app.Fragment) need to be replaced with the support library
version (android.support.v4.app.Fragment). Consequently, the same must
be done for LoaderManager and Loader, and a few other changes are
needed. Yay!