Package renaming from long ago
It was a while ago — certainly I don't remember it now, but apparently I
decided to change all the package names. This probably had to do with
the fact that Ion was going to replace Iodine, which at this point it
already has. So I might as well delete the entire iodine subpackage.